WIP tests against Django 1.6b1, some tests failing with Python3
