+++ CHANGES
@@ -1,7 +1,7 @@
 Version 0.7.3
 -------------
 * Fix for test settings leaking onto global settings: LANGUAGES was overridden and not set back (Issue #81 - Thanks @zsoldosp)
-
+* Test against Django 1.6b1
 
 Version 0.7.2
 -------------
