+++ setup.py
@@ -18,6 +18,9 @@ setup(
         'Topic :: Software Development :: Localization',
         'Topic :: Software Development :: Internationalization',
         'Framework :: Django',
+        'Programming Language :: Python :: 2.6',
+        'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 3.3',
     ],
     include_package_data=True,
     zip_safe=False,
